Migrate logging functionality to shared library
