+++ libraries/networking/src/HifiSockAddr.h
@@ -91,9 +91,6 @@ namespace std {
     };
 }
 
-
-QHostAddress getGuessedLocalAddress();
-
 Q_DECLARE_METATYPE(HifiSockAddr);
 
 #endif // hifi_HifiSockAddr_h
