+++ libraries/shared/src/shared/AbstractLoggerInterface.h
