+++ libraries/shared/src/shared/NetworkUtils.h
@@ -1,0 +1,17 @@
+//
+//  Created by Bradley Austin Davis on 2016/09/20
+//  Copyright 2013-2016 High Fidelity, Inc.
+//
+//  Distributed under the Apache License, Version 2.0.
+//  See the accompanying file LICENSE or http://www.apache.org/licenses/LICENSE-2.0.html
+//
+
+#pragma once
+#ifndef hifi_shared_NetworkUtils_h
+#define hifi_shared_NetworkUtils_h
+
+#include <QtNetwork/QHostAddress>
+
+QHostAddress getGuessedLocalAddress();
+
+#endif // hifi_shared_NetworkUtils_h
